+++ engine/flink/api/src/main/scala/pl/touk/nussknacker/engine/flink/api/compat/FlinkCompatibilityProvider.scala
@@ -1,0 +1,5 @@
+package pl.touk.nussknacker.engine.flink.api.compat
+
+trait FlinkCompatibilityProvider {
+
+}
