+++ engine/flink/process/src/main/scala/pl/touk/nussknacker/engine/process/FlinkCompatibilityProvider.scala
@@ -1,0 +1,11 @@
+package pl.touk.nussknacker.engine.process
+
+import com.typesafe.config.Config
+import pl.touk.nussknacker.engine.process.registrar.StreamExecutionEnvPreparer
+
+trait FlinkCompatibilityProvider {
+
+  def createExecutionEnvPreparer(config: Config,
+                                 executionConfigPreparer: ExecutionConfigPreparer, useDiskState: Boolean): StreamExecutionEnvPreparer
+
+}
